# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reservasi Ruangan DinuSpace",
        "description": "Dokumentasi API untuk sistem reservasi ruangan Gedung H.",
        "contact": {
            "name": "Tim Developer",
            "email": "support@dinus.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
...
        "/admin/rooms": {
            "get": {
                "tags": [
                    "Admin - Manajemen Ruangan"
                ],
                "summary": "Melihat Ruangan",
                "description": "Endpoint untuk melihat data ruangan",
                "operationId": "bebf0b76e24db97701e4688b00b7928c",
                "responses": {
                    "200": {
                        "description": "Data Berhasil di tampilkan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Kesalahan Server.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "title": "Respons sukses standar",
                "properties": {
                    "data": {
                        "description": "Data hasil request",
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "status_code": {
                                "type": "integer",
                                "example": 200
                            },
                            "success": {
                                "type": "boolean",
                                "example": true
                            },
                            "message": {
                                "type": "string",
                                "example": "Berhasil mengambil data."
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Respons error standar",
                "properties": {
                    "data": {
                        "type": "null"
                    },
                    "meta": {
                        "properties": {
                            "status_code": {
                                "type": "integer",
                                "example": 400
                            },
                            "success": {
                                "type": "boolean",
                                "example": false
                            },
                            "message": {
                                "type": "string",
                                "example": "Kredensial tidak valid."
                            },
                            "errors": {
                                "type": "string",
                                "example": "Exception Error : Unauthorized"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}